+++ src/app/dashboard/dashboard.module.ts
@@ -6,10 +6,10 @@ import { HomeComponent } from './home/home.component';
 import { NgApexchartsModule } from 'ng-apexcharts';
 import { ProgressoLimpezaChartComponent } from './progresso-limpeza-chart/progresso-limpeza-chart.component';
 import { HistoricoOcorrenciaChartComponent } from './historico-ocorrencia-chart/historico-ocorrencia-chart.component';
-import { MetaFuncionariosChartComponent } from './meta-funcionarios-chart/meta-funcionarios-chart.component';
+import { ProgressoLimpezaFuncionarioChartComponent } from './progresso-limpeza-funcionario-chart/progresso-limpeza-funcionario-chart.component';
 
 @NgModule({
-  declarations: [HomeComponent, ProgressoLimpezaChartComponent, HistoricoOcorrenciaChartComponent, MetaFuncionariosChartComponent],
+  declarations: [HomeComponent, ProgressoLimpezaChartComponent, HistoricoOcorrenciaChartComponent, ProgressoLimpezaFuncionarioChartComponent],
   imports: [CommonModule, DashboardRoutingModule, NgApexchartsModule],
 })
 export class DashboardModule {}
