--- src/app/dashboard/meta-funcionarios-chart/meta-funcionarios-chart.component.ts
@@ -1,190 +1,0 @@
-import { Component, ViewChild, OnInit } from "@angular/core";
-import {
-  ApexAxisChartSeries,
-  ApexChart,
-  ChartComponent,
-  ApexDataLabels,
-  ApexPlotOptions,
-  ApexLegend
-} from "ng-apexcharts";
-
-export type ChartOptions = {
-  series: ApexAxisChartSeries;
-  chart: ApexChart;
-  dataLabels: ApexDataLabels;
-  plotOptions: ApexPlotOptions;
-  legend: ApexLegend;
-  colors: string[];
-};
-
-@Component({
-  selector: 'app-meta-funcionarios-chart',
-  templateUrl: './meta-funcionarios-chart.component.html',
-  styleUrls: ['./meta-funcionarios-chart.component.scss']
-})
-export class MetaFuncionariosChartComponent implements OnInit {
-  @ViewChild("chart") chart: ChartComponent;
-  public chartOptions: Partial<ChartOptions>;
-  private observer: MutationObserver;
-
-  constructor() {}
-
-  ngOnInit() {
-    this.initializeChartOptions();
-    this.setupThemeObserver();
-  }
-
-  ngOnDestroy() {
-    this.observer.disconnect();
-  }
-
-  initializeChartOptions() {
-    const isDarkMode = document.documentElement.classList.contains('dark'); // Check if dark mode is enabled on <html> element
-    const textColor = isDarkMode ? '#FFFFFF' : '#000000'; // Choose text color based on theme
-    const goalStrokeColor = isDarkMode ? '#FFFFFF' : '#775DD0'; // Adjust based on your preference
-    const backgroundColor = isDarkMode ? '#4B5563' : '#F9FAFB'; // Adjust based on your preference
-
-    this.chartOptions = {
-      series: [
-        {
-          name: "Actual",
-          data: [
-            {
-              x: "2011",
-              y: 1292,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 1400,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2012",
-              y: 4432,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 5400,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2013",
-              y: 5423,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 5200,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2014",
-              y: 6653,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 6500,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2015",
-              y: 8133,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 6600,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2016",
-              y: 7132,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 7500,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2017",
-              y: 7332,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 8700,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            },
-            {
-              x: "2018",
-              y: 6553,
-              goals: [
-                {
-                  name: "Expected",
-                  value: 7300,
-                  strokeWidth: 5,
-                  strokeColor: goalStrokeColor
-                }
-              ]
-            }
-          ]
-        }
-      ],
-      chart: {
-        height: 350,
-        type: "bar",
-        foreColor: textColor,
-        background: backgroundColor
-      },
-      plotOptions: {
-        bar: {
-          columnWidth: "60%"
-        }
-      },
-      colors: ["#00E396"],
-      dataLabels: {
-        enabled: false
-      },
-      legend: {
-        show: true,
-        showForSingleSeries: true,
-        customLegendItems: ["Actual", "Expected"],
-        markers: {
-          fillColors: ["#00E396", goalStrokeColor]
-        }
-      }
-    };
-  }
-
-  setupThemeObserver() {
-    this.observer = new MutationObserver((mutations) => {
-      mutations.forEach((mutation) => {
-        if (mutation.attributeName === 'class') {
-          this.initializeChartOptions(); // Re-initialize chart options on theme change
-        }
-      });
-    });
-
-    this.observer.observe(document.documentElement, {
-      attributes: true
-    });
-  }
-}
