+++ src/app/dashboard/progresso-limpeza-chart/progresso-limpeza-chart.component.html
@@ -3,7 +3,7 @@
   <div class="flex justify-between mb-3">
     <div class="flex items-center">
       <div class="flex justify-center items-center">
-        <h5 class="text-xl font-bold leading-none text-gray-900 dark:text-white pe-1">Progresso de Limpeza do Dia</h5>
+        <h5 class="text-xl font-bold leading-none font-roboto text-gray-900 dark:text-white pe-1">Progresso de Limpeza do Dia</h5>
       </div>
     </div>
   </div>
