+++ src/app/dashboard/progresso-limpeza-funcionario-chart/progresso-limpeza-funcionario-chart.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ProgressoLimpezaFuncionarioChartComponent } from './progresso-limpeza-funcionario-chart.component';
+
+describe('ProgressoLimpezaFuncionarioChartComponent', () => {
+  let component: ProgressoLimpezaFuncionarioChartComponent;
+  let fixture: ComponentFixture<ProgressoLimpezaFuncionarioChartComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [ProgressoLimpezaFuncionarioChartComponent]
+    });
+    fixture = TestBed.createComponent(ProgressoLimpezaFuncionarioChartComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
